+++ src/components/Modals/ModalButton/style.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+import { globalColor } from '../../../GlobalStyle';
+
+export const ModalSubmitBtn = styled.button`
+  width: 14.69rem;
+  height: 2.625rem;
+  background-color: ${globalColor};
+  border-radius: 1rem;
+  border: none;
+  box-shadow: 0 0.1875rem 0.62rem 0 rgba(0, 0, 0, 0.27);
+  font-size: 0.9375rem;
+  color: #fff;
+  cursor: pointer;
+  margin-bottom: 3rem;
+`;
