+++ src/components/Modals/ModalWrapper/style.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
-import { globalColor } from '../../GlobalStyle';
-import ModalCloseImg from '../assets/icon/cross-mark.png';
+import { globalColor } from '../../../GlobalStyle';
 
 export const ModalOverlay = styled.div`
   position: fixed;
@@ -62,17 +61,4 @@ export const ModalInput = styled.input`
   line-height: normal;
   padding-left: 1.125rem;
   font-size: 1rem;
-`;
-
-export const ModalSubmitBtn = styled.button`
-  width: 14.69rem;
-  height: 2.625rem;
-  background-color: ${globalColor};
-  border-radius: 1rem;
-  border: none;
-  box-shadow: 0 0.1875rem 0.62rem 0 rgba(0, 0, 0, 0.27);
-  font-size: 0.9375rem;
-  color: #fff;
-  cursor: pointer;
-  margin-bottom: 3rem;
 `;
